splitting up integer test suite to have the same structure as the other number systems
